updated and retested error handling
