+++ Backend/WebApi-PointMap/WebApi-PointMap.csproj
@@ -112,12 +112,7 @@
     <Compile Include="Controllers\SessionController.cs" />
     <Compile Include="Controllers\TesterController.cs" />
     <Compile Include="Controllers\UserController.cs" />
-    <Compile Include="ErrorHandling\AuthorizationErrorHandler.cs" />
-    <Compile Include="ErrorHandling\GeneralErrorHandler.cs" />
-    <Compile Include="ErrorHandling\GlobalErrorLogger.cs" />
-    <Compile Include="ErrorHandling\DatabaseErrorHandler.cs" />
-    <Compile Include="ErrorHandling\HttpErrorHandler.cs" />
-    <Compile Include="ErrorHandling\PointErrorHandler.cs" />
+    <Compile Include="ErrorHandling\ErrorHandler.cs" />
     <Compile Include="Filters\GlobalExceptionFilter.cs" />
     <Compile Include="Controllers\UserManagementController.cs" />
     <Compile Include="Global.asax.cs">
